requirements.txt
alive-progress==3.1.5
apache-superset==4.0.2
dbt-core>=1.7
dbt-duckdb>=1.7
dlt[duckdb]>=0.5.2
duckdb-engine==0.13.1
enlighten==1.12.4
invoke==2.2.0
python-dotenv==1.0.1
sqlfluff==3.1.0
streamlit==1.37.1
